dealer-email-finder/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
openpyxl>=3.1.0
ddgs>=6.0.0
dnspython>=2.4.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
sse-starlette>=2.0.0
pandas>=2.0.0
email-validator>=2.0.0
playwright>=1.40.0
